# Reduce lock contentions in TabletsWriter