--- a/be/src/runtime/load_channel_mgr.cpp
+++ b/be/src/runtime/load_channel_mgr.cpp
@@ -23,12 +23,11 @@
 
 #include <memory>
 
+#include "common/closure_guard.h"
 #include "gutil/strings/substitute.h"
-#include "runtime/exec_env.h"
 #include "runtime/load_channel.h"
 #include "runtime/mem_tracker.h"
 #include "service/backend_options.h"
-#include "storage/lru_cache.h"
 #include "util/starrocks_metrics.h"
 #include "util/stopwatch.hpp"
 #include "util/thread.h"
@@ -58,7 +57,6 @@ LoadChannelMgr::LoadChannelMgr() : _is_stopped(false) {
         std::lock_guard<std::mutex> l(_lock);
         return _load_channels.size();
     });
-    _lastest_success_channel = new_lru_cache(1024);
 }
 
 LoadChannelMgr::~LoadChannelMgr() {
@@ -66,7 +64,6 @@ LoadChannelMgr::~LoadChannelMgr() {
     if (_load_channels_clean_thread.joinable()) {
         _load_channels_clean_thread.join();
     }
-    delete _lastest_success_channel;
 }
 
 Status LoadChannelMgr::init(MemTracker* mem_tracker) {
@@ -75,94 +72,52 @@ Status LoadChannelMgr::init(MemTracker* mem_tracker) {
     return Status::OK();
 }
 
-Status LoadChannelMgr::open(const PTabletWriterOpenRequest& params) {
-    UniqueId load_id(params.id());
-    std::shared_ptr<LoadChannel> channel;
+void LoadChannelMgr::open(brpc::Controller* cntl, const PTabletWriterOpenRequest* request,
+                          PTabletWriterOpenResult* response, google::protobuf::Closure* done) {
+    ClosureGuard done_guard(done);
+    UniqueId load_id(request->id());
+    scoped_refptr<LoadChannel> channel;
     {
         std::lock_guard<std::mutex> l(_lock);
         auto it = _load_channels.find(load_id);
         if (it != _load_channels.end()) {
             channel = it->second;
         } else {
-            int64_t mem_limit_in_req = params.has_load_mem_limit() ? params.load_mem_limit() : -1;
+            int64_t mem_limit_in_req = request->has_load_mem_limit() ? request->load_mem_limit() : -1;
             int64_t job_max_memory = calc_job_max_load_memory(mem_limit_in_req, _mem_tracker->limit());
 
-            int64_t timeout_in_req_s = params.has_load_channel_timeout_s() ? params.load_channel_timeout_s() : -1;
+            int64_t timeout_in_req_s = request->has_load_channel_timeout_s() ? request->load_channel_timeout_s() : -1;
             int64_t job_timeout_s = calc_job_timeout_s(timeout_in_req_s);
             auto job_mem_tracker = std::make_unique<MemTracker>(job_max_memory, load_id.to_string(), _mem_tracker);
 
-            channel.reset(new LoadChannel(load_id, job_timeout_s, std::move(job_mem_tracker)));
+            channel.reset(new LoadChannel(this, load_id, job_timeout_s, std::move(job_mem_tracker)));
             _load_channels.insert({load_id, channel});
         }
     }
-
-    RETURN_IF_ERROR(channel->open(params));
-    return Status::OK();
+    channel->open(cntl, request, response, done_guard.release());
 }
 
-static void dummy_deleter(const CacheKey& key, void* value) {}
-
-Status LoadChannelMgr::add_chunk(const PTabletWriterAddChunkRequest& request,
-                                 google::protobuf::RepeatedPtrField<PTabletInfo>* tablet_vec,
-                                 int64_t* wait_lock_time_ns) {
-    UniqueId load_id(request.id());
-    // 1. get load channel
-    std::shared_ptr<LoadChannel> channel;
-    {
-        std::lock_guard<std::mutex> l(_lock);
-        auto it = _load_channels.find(load_id);
-        if (it == _load_channels.end()) {
-            auto* handle = _lastest_success_channel->lookup(load_id.to_string());
-            // success only when eos be true
-            if (handle != nullptr) {
-                _lastest_success_channel->release(handle);
-                if (request.has_eos() && request.eos()) {
-                    return Status::OK();
-                }
-            }
-            return Status::InternalError(
-                    strings::Substitute("fail to add batch in load channel. unknown load_id=$0", load_id.to_string()));
-        }
-        channel = it->second;
+void LoadChannelMgr::add_chunk(brpc::Controller* cntl, const PTabletWriterAddChunkRequest* request,
+                               PTabletWriterAddBatchResult* response, google::protobuf::Closure* done) {
+    VLOG(2) << "Current memory usage=" << _mem_tracker->consumption() << " limit=" << _mem_tracker->limit();
+    ClosureGuard done_guard(done);
+    UniqueId load_id(request->id());
+    auto channel = _find_load_channel(load_id);
+    if (channel != nullptr) {
+        channel->add_chunk(cntl, request, response, done_guard.release());
+    } else {
+        response->mutable_status()->set_status_code(TStatusCode::INTERNAL_ERROR);
+        response->mutable_status()->add_error_msgs("no associated load channel");
     }
-
-    // 3. add batch to load channel
-    // batch may not exist in request(eg: eos request without batch),
-    // this case will be handled in load channel's add batch method.
-    RETURN_IF_ERROR(channel->add_chunk(request, tablet_vec));
-
-    // 4. handle finish
-    if (channel->is_finished()) {
-        LOG(INFO) << "Removing finished load channel load id=" << load_id;
-        {
-            std::lock_guard<std::mutex> l(_lock);
-            _load_channels.erase(load_id);
-            auto* handle = _lastest_success_channel->insert(load_id.to_string(), nullptr, 1, dummy_deleter);
-            _lastest_success_channel->release(handle);
-        }
-        VLOG(1) << "Removed load channel load id=" << load_id;
-    }
-
-    return Status::OK();
 }
 
-Status LoadChannelMgr::cancel(const PTabletWriterCancelRequest& params) {
-    UniqueId load_id(params.id());
-    std::shared_ptr<LoadChannel> cancelled_channel;
-    {
-        std::lock_guard<std::mutex> l(_lock);
-        if (_load_channels.find(load_id) != _load_channels.end()) {
-            cancelled_channel = _load_channels[load_id];
-            _load_channels.erase(load_id);
-        }
+void LoadChannelMgr::cancel(brpc::Controller* cntl, const PTabletWriterCancelRequest* request,
+                            PTabletWriterCancelResult* response, google::protobuf::Closure* done) {
+    ClosureGuard done_guard(done);
+    UniqueId load_id(request->id());
+    if (auto channel = remove_load_channel(load_id); channel != nullptr) {
+        channel->cancel();
     }
-
-    if (cancelled_channel != nullptr) {
-        cancelled_channel->cancel();
-        LOG(INFO) << "Cancelled load channel load id=" << load_id;
-    }
-
-    return Status::OK();
 }
 
 Status LoadChannelMgr::_start_bg_worker() {
@@ -186,14 +141,13 @@ Status LoadChannelMgr::_start_bg_worker() {
 }
 
 Status LoadChannelMgr::_start_load_channels_clean() {
-    std::vector<std::shared_ptr<LoadChannel>> timeout_channels;
+    std::vector<scoped_refptr<LoadChannel>> timeout_channels;
+
     time_t now = time(nullptr);
     {
         std::lock_guard<std::mutex> l(_lock);
-        VLOG(1) << "there are " << _load_channels.size() << " running load channels";
         for (auto it = _load_channels.begin(); it != _load_channels.end(); /**/) {
-            time_t last_updated_time = it->second->last_updated_time();
-            if (difftime(now, last_updated_time) >= it->second->timeout()) {
+            if (difftime(now, it->second->last_updated_time()) >= it->second->timeout()) {
                 timeout_channels.emplace_back(std::move(it->second));
                 it = _load_channels.erase(it);
             } else {
@@ -216,6 +170,22 @@ Status LoadChannelMgr::_start_load_channels_clean() {
               << " current=" << _mem_tracker->consumption() << " peak=" << _mem_tracker->peak_consumption();
 
     return Status::OK();
+}
+
+scoped_refptr<LoadChannel> LoadChannelMgr::_find_load_channel(const UniqueId& load_id) {
+    std::lock_guard l(_lock);
+    auto it = _load_channels.find(load_id);
+    return (it != _load_channels.end()) ? it->second : nullptr;
+}
+
+scoped_refptr<LoadChannel> LoadChannelMgr::remove_load_channel(const UniqueId& load_id) {
+    std::lock_guard l(_lock);
+    if (auto it = _load_channels.find(load_id); it != _load_channels.end()) {
+        auto ret = it->second;
+        _load_channels.erase(it);
+        return ret;
+    }
+    return nullptr;
 }
 
 } // namespace starrocks