--- a/be/src/runtime/load_channel_mgr.h
+++ b/be/src/runtime/load_channel_mgr.h
@@ -27,17 +27,23 @@
 #include <thread>
 #include <unordered_map>
 
-#include "common/status.h"
+#include "common/statusor.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
+#include "runtime/load_channel.h"
 #include "runtime/tablets_channel.h"
+#include "util/blocking_queue.hpp"
+#include "util/threadpool.h"
 #include "util/uid_util.h"
+
+namespace brpc {
+class Controller;
+}
 
 namespace starrocks {
 
 class Cache;
-class LoadChannel;
 
 // LoadChannelMgr -> LoadChannel -> TabletsChannel -> DeltaWriter
 // All dispatched load data for this backend is routed from this class
@@ -48,23 +54,26 @@ public:
 
     Status init(MemTracker* mem_tracker);
 
-    // open a new load channel if not exist
-    Status open(const PTabletWriterOpenRequest& request);
+    void open(brpc::Controller* cntl, const PTabletWriterOpenRequest* request, PTabletWriterOpenResult* response,
+              google::protobuf::Closure* done);
 
-    Status add_chunk(const PTabletWriterAddChunkRequest& request,
-                     google::protobuf::RepeatedPtrField<PTabletInfo>* tablet_vec, int64_t* wait_lock_time_ns);
+    void add_chunk(brpc::Controller* cntl, const PTabletWriterAddChunkRequest* request,
+                   PTabletWriterAddBatchResult* response, google::protobuf::Closure* done);
 
-    // cancel all tablet stream for 'load_id' load
-    Status cancel(const PTabletWriterCancelRequest& request);
+    void cancel(brpc::Controller* cntl, const PTabletWriterCancelRequest* request, PTabletWriterCancelResult* response,
+                google::protobuf::Closure* done);
+
+    scoped_refptr<LoadChannel> remove_load_channel(const UniqueId& load_id);
 
 private:
     Status _start_bg_worker();
 
+    scoped_refptr<LoadChannel> _find_load_channel(const UniqueId& load_id);
+
     // lock protect the load channel map
     std::mutex _lock;
     // load id -> load channel
-    std::unordered_map<UniqueId, std::shared_ptr<LoadChannel>> _load_channels;
-    Cache* _lastest_success_channel = nullptr;
+    std::unordered_map<UniqueId, scoped_refptr<LoadChannel>> _load_channels;
 
     // check the total load mem consumption of this Backend
     MemTracker* _mem_tracker = nullptr;